Safari-609.1.11.1/LayoutTests/platform/ios/ios/fast/coordinates/mouse-event-zoomed.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
<script src="resources/helpers.js"></script>
<style>* { margin: 0; padding: 0; }</style>
</head>
<body>
<div id="box" style="background:blue;width:100px;height:100px;position:absolute;top:100px;left:100px"></div>
<p id="description"></p>
<div id="console"></div>
<!-- Filler to allow scrolling -->
<div style="width:1500px;height:1500px;"></div>
<script>
description("This tests mouse event coordinates at scale 2 and scrolled slightly.");

var box = document.getElementById('box');
box.onmousedown = function(event) { debug('mousedown'); verifyCoordinates(event); }
box.onmouseup = function(event) { debug('mouseup'); verifyCoordinates(event); }
box.onclick = function(event) { debug('click'); verifyCoordinates(event); endTest(); }

const scale = 2;
const panX = 10;
const panY = 10;

function sendMouseSequence() {
    if (!window.eventSender) {
         debug("This test requires DumpRenderTree to send mouse events.");
         return;
    }

    setExpectedValues(125+panX, 175+panY, 125, 175);
    eventSender.mouseMoveTo(125*scale, 175*scale);
    eventSender.mouseDown();
    setExpectedValues(175+panX, 125+panY, 175, 125);
    eventSender.mouseMoveTo(175*scale, 125*scale);
    eventSender.mouseUp();
}

setInitialScaleAndPanBy(scale, panX, panY);
sendMouseSequence();

var successfullyParsed = true;
</script>
</body>
</html>
