Safari-609.1.11.1/LayoutTests/platform/ios/ios/fast/coordinates/mouse-event-zoomed-expected.txt

This tests mouse event coordinates at scale 2 and scrolled slightly.
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
mousedown
PASS eventOrTouch.pageX is 135
PASS eventOrTouch.pageY is 185
PASS eventOrTouch.clientX is 125
PASS eventOrTouch.clientY is 175
mouseup
PASS eventOrTouch.pageX is 185
PASS eventOrTouch.pageY is 135
PASS eventOrTouch.clientX is 175
PASS eventOrTouch.clientY is 125
click
PASS eventOrTouch.pageX is 185
PASS eventOrTouch.pageY is 135
PASS eventOrTouch.clientX is 175
PASS eventOrTouch.clientY is 125
PASS successfullyParsed is true

TEST COMPLETE
